Changed the look for the disclosure group
